requirements.txt
numpy==2.2.1
pandas==2.2.3
Pillow==11.1.0
streamlit==1.41.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
